feature-hasPermissions

+++ app/Http/Controllers/CategoryController.php
@@ -18,6 +18,7 @@ class CategoryController extends Controller
 
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Category::class);
         $items = $this->categoryService->all($request);
         // dd($items);
         return view('admin.categories.index', compact('items'));
@@ -25,6 +26,7 @@ class CategoryController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Category::class);
         return view('admin.categories.create');
     }
 
@@ -37,6 +39,7 @@ class CategoryController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('update', Category::class);
         $items = $this->categoryService->find($id);
         return view('admin.categories.edit', compact('items'));
     }
@@ -50,6 +53,7 @@ class CategoryController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('delete', Category::class);
         $items = $this->categoryService->delete($id);
         return redirect()->route('categories.index');
     }
@@ -61,11 +65,13 @@ class CategoryController extends Controller
     }
     public function restore($id)
     {
+        $this->authorize('restore', Category::class);
         $items = $this->categoryService->restore($id);
         return redirect()->route('categories.index');
     }
     public function deleteforever($id)
     {
+        $this->authorize('deleteforever', Category::class);
         $items = $this->categoryService->deleteforever($id);
         return redirect()->route('categories.index');
     }
